+++ src/common/scene/Camera.cpp
@@ -86,7 +86,7 @@ void Camera::handleInput(GLFWwindow *window, float delta)
 	setView(position, position + front, up);
 }
 
-void Camera::handleMouseInput(double x, double y)
+void Camera::handleMouseInput(float x, float y)
 {
 	if (inititalMouse)
 	{
